refactor(core): move implementation detail header and types
