+++ includes/acl/compression/impl/quantize_track_impl.h
@@ -26,7 +26,7 @@
 
 #include "acl/core/impl/compiler_utils.h"
 #include "acl/core/track_types.h"
-#include "acl/core/variable_bit_rates.h"
+#include "acl/core/impl/variable_bit_rates.h"
 #include "acl/compression/impl/track_list_context.h"
 
 #include <rtm/mask4i.h>
