+++ includes/acl/compression/impl/sample_streams.h
@@ -29,7 +29,7 @@
 #include "acl/core/error.h"
 #include "acl/core/time_utils.h"
 #include "acl/core/track_formats.h"
-#include "acl/core/variable_bit_rates.h"
+#include "acl/core/impl/variable_bit_rates.h"
 #include "acl/math/quat_packing.h"
 #include "acl/math/vector4_packing.h"
 #include "acl/compression/impl/track_stream.h"
