+++ includes/acl/compression/impl/write_stats.h
@@ -28,7 +28,7 @@
 
 #include "acl/core/time_utils.h"
 #include "acl/core/track_formats.h"
-#include "acl/core/variable_bit_rates.h"
+#include "acl/core/impl/variable_bit_rates.h"
 #include "acl/core/impl/compiler_utils.h"
 #include "acl/compression/transform_error_metrics.h"
 #include "acl/compression/track_error.h"
