+++ includes/acl/compression/impl/write_stream_data.h
@@ -28,7 +28,7 @@
 #include "acl/core/iallocator.h"
 #include "acl/core/error.h"
 #include "acl/core/track_formats.h"
-#include "acl/core/variable_bit_rates.h"
+#include "acl/core/impl/variable_bit_rates.h"
 #include "acl/compression/impl/animated_track_utils.h"
 #include "acl/compression/impl/clip_context.h"
