+++ includes/acl/compression/impl/write_track_data_impl.h
@@ -24,9 +24,9 @@
 // SOFTWARE.
 ////////////////////////////////////////////////////////////////////////////////
 
-#include "acl/core/variable_bit_rates.h"
 #include "acl/core/impl/compiler_utils.h"
 #include "acl/core/impl/compressed_headers.h"
+#include "acl/core/impl/variable_bit_rates.h"
 #include "acl/compression/impl/track_list_context.h"
 
 #include <rtm/vector4f.h>
