+++ includes/acl/decompression/impl/scalar_track_decompression.h
@@ -28,8 +28,8 @@
 #include "acl/core/compressed_tracks_version.h"
 #include "acl/core/interpolation_utils.h"
 #include "acl/core/track_writer.h"
-#include "acl/core/variable_bit_rates.h"
 #include "acl/core/impl/compiler_utils.h"
+#include "acl/core/impl/variable_bit_rates.h"
 #include "acl/decompression/database/database.h"
 #include "acl/math/scalar_packing.h"
 #include "acl/math/vector4_packing.h"
