+++ includes/acl/decompression/impl/transform_track_decompression.h
@@ -32,8 +32,8 @@
 #include "acl/core/range_reduction_types.h"
 #include "acl/core/track_formats.h"
 #include "acl/core/track_writer.h"
-#include "acl/core/variable_bit_rates.h"
 #include "acl/core/impl/compiler_utils.h"
+#include "acl/core/impl/variable_bit_rates.h"
 #include "acl/decompression/database/database.h"
 #include "acl/decompression/impl/transform_animated_track_cache.h"
 #include "acl/decompression/impl/transform_constant_track_cache.h"
