+++ tests/sources/math/test_scalar_packing.cpp
@@ -24,7 +24,7 @@
 
 #include <catch2/catch.hpp>
 
-#include <acl/core/variable_bit_rates.h>
+#include <acl/core/impl/variable_bit_rates.h>
 #include <acl/math/scalar_packing.h>
 #include <acl/math/vector4_packing.h>
 
@@ -124,9 +124,9 @@ TEST_CASE("unpack_scalarf_uXX_unsafe", "[math][scalar][packing]")
 		if (!scalar_near_equal(vector_get_x(vec0), scalar_cast(scalar1), 1.0E-6F))
 			num_errors++;
 
-		for (uint8_t bit_rate = 1; bit_rate < k_highest_bit_rate; ++bit_rate)
+		for (uint8_t bit_rate = 1; bit_rate < acl_impl::k_highest_bit_rate; ++bit_rate)
 		{
-			uint32_t num_bits = get_num_bits_at_bit_rate(bit_rate);
+			uint32_t num_bits = acl_impl::get_num_bits_at_bit_rate(bit_rate);
 			uint32_t max_value = (1 << num_bits) - 1;
 			for (uint32_t value = 0; value <= max_value; ++value)
 			{
